feat: Allow update of PGPSecret and PrivateKey
Updated the python code that backs the custom resource implementations for the
PGPSecret and PrivateKey constructs so they allow updating the KMS encryption
keys used by AWS SecretsManager without necessarily re-generating new secrets.

Also, added a `resourceVersion` has to the properties so the resource handler is
called again if the implementation code changed.

BREAKING CHANGE: this also changes the API of the PGPSecret and
CodeSigningCertificate constructs to offer a consistent API for accessing the name
and ARNs of the secret and parameters associated with the secrets, through the
`ICredentialPair` interface.