fix: Stop surfacing and using secret VersionIds
The secrets VersionId field can be used to reference a particular instance of a secret
however this makes migrating secrets to a new location (by replacing the content with
a new value) impossible, and also doesn't work great with automatic secret rotation.
Hene, removing the option to even access it.