feat: Create OpenPGP Public Key parameter using SSM resource
Stop creating the parameter directly from the custom resource
implementation as the SSM parameters are not "importable" (the value
reference is extremely awkward to obtain). Instead, return the public
key as part of the custom resource response, and create an
`AWS::SSM::Parameter` resource in the template.

This allows updating the secrets to return `smm.IParameter` and
`secretsManager.ISecretString` instead of returning the respective ARNs.