feat: Support Change Control Policies
Storing an iCal document in an S3 bucket, it can be used to define
windows of time during which promotions to the `Publish` stage will be
disabled. The event summary will be used as the `reason` for the
transition disabling operation.

The enforcement lambda is invoked periodically (by default, every 15
minutes), and the window closes a set amount of time (1 hour) before the
next event starts (such that publishing does not normally happen during
a window state transition).

Fixes awslabs/cdk-ops#231