feat: multiple artifacts for GitHub and Shellable
Add multiple output artifact support for Shellable (including an initial
attempt at a BuildSpec helper as requested in awslabs/aws-cdk#2039).

Add multiple input artifact support for the GitHub publisher. The
tag and release notes will be taken from the primary artifact, the other
artifacts will be attached to the same release.

This is in preparation of publishing downloadable reference docs to
GitHub.