feat: multiple artifacts for GitHub and Shellable  (#90)
Add multiple output artifact support for Shellable (including an initial
attempt at a BuildSpec helper as requested in awslabs/aws-cdk#2039).

Add multiple input artifact support for the GitHub publisher. The
tag and release notes will be taken from the primary artifact, the other
artifacts will be attached to the same release.
The CHANGELOG will only be taken from the primary artifact.

This is in preparation of publishing downloadable reference docs to
GitHub.

Add a global `dryRun` option to Pipeline to globally disable publishing.