feat: auto-build (with public logs)
If the `autoBuild` flag is set, a CodeBuild project
will be created and configured to automatically build
any commits to the repo, including GitHub pull requests.

If the `publicLogs` flag is also set, build logs will be published
to S3 and also a comment will automatically be posted on the GitHub
PR with a link.

Fixes #42