feat: configure buildspec for AutoBuild project
Allow a different buildspec to be specified for the AutoBuild project.
This enables PR builds to use a different buildspec file from the
CodeBuild project in the pipeline.

Motivation
The aws-cdk package spends a significant amount of time (~20 minutes) in
the post-build step, where the build artifacts are packed.
Specifying a separate buildspec.yml that has no post-build step will
speed up PR builds.