feat: migrate library to General Availability CDK version
This updates the CDK libraries to a post-1.0.0 version.
It contains only the minimal amount of changes needed for the code to compile & pass tests.
Further changes, needed to accommodate the changes in CDK conventions that happened since this library was created, will be coming later.

This is an empty commit, because I accidentally merged the PR with the actual changes from the GitHub UI
without any message (and this is a problem, as we're using conventional commits).

BREAKING CHANGE: AutoBuildOptions.buildSpec is now of type codebuild.BuildSpec
BREAKING CHANGE: CanaryProps.scheduleExpression is now of type events.Schedule, and was renamed to 'schedule'
BREAKING CHANGE: ChangeControllerProps.scheduleExpression is now of type events.Schedule, and was renamed to 'schedule'
BREAKING CHANGE: RsaPrivateKeySecretProps.deletionPolicy has been renamed to 'removalPolicy'
BREAKING CHANGE: PipelineProps.buildSpec is now of type codebuild.BuildSpec
BREAKING CHANGE: GitHubRepo.tokenParameterName is now of type cdk.SecretValue, and was renamed to 'token'
BREAKING CHANGE: ShellableOptions.alarmPeriodSec is now of type cdk.Duration, and was renamed to 'alarmPeriod'