feat: upgrade Custom Resources to use Lambda's Node10 runtime (#226)
feat: upgrade Custom Resources to use Lambda's Node10 runtime

Our custom resources for signing and key management shell out to OpenSSL and GPG programs. Those binaries no longer ship with Lambda's Node runtimes later than 8. This changes the Lambda functions backing the custom resources to include layers that add those binaries. In the case of OpenSSL, just the CLI binary is enough, as the libraries are present in the runtime; for GPG, however, we also need to include a bunch of shared libraries, and the `gpg-agent` binary, required for managing private keys with GPG 2.0.

All binaries were obtained by downloading them from an EC2 instance running Amazon Linux 2,
and zipping them up to be used in a layer.

The invocation of the commands had to change to refer to the binaries in `/opt/`, which is where the Lambda layers wind up, and some additional required options were added because of the way Lambda gives read-only permissions to the underlying filesystem.

Fixes #220

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.