fix: public build logs are broken
The public build logs Serverless App retrieves the GitHub Token from the
CodeBuild project it's monitoring. For some reason (maybe because we
switched to Secrets Manager secrets) that token can't be retrieved
anymore and the posting would fail.

Also, the specific fields that the serverless app was reading have
been loudly deprecated according to the CodeBuild docs:

https://docs.aws.amazon.com/codebuild/latest/APIReference/API_ProjectSource.html#CodeBuild-Type-ProjectSource-auth

Update to a newer version of the serverless app in which this
token is an explicit parameter.