feat(bump): create pull request for bumps
Add support for automatically creating pull requests for bumps.

BREAKING CHANGE: This requires the GitHub token to be available during bump. However, currently we model the GitHub token as a SecretValue, which resolves during deployment (by CloudFormation). This is not secure because it will pass the secret itself as environment variables. To address this, we changed `token` to `tokenSecretArn` and resolve it as needed (either during deployment if it’s required for a CFN definition or during runtime by extracting it from secrets manager).