fix: Grant env var access in shellable when needed (#253)
fix: Grant env var access in shellable when needed

Calls `grantRead` methods for secrets and SSM parameters passed as env
vars to shellable. Originally meant for consumers to handle on their end
but constructs don't necessarily expose the roles to be given access.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.