chore(pipeline): configure name and expose build project
Exposing the CodeBuild project in the pipeline will enable use cases
around creating monitoring resources. As an example, the build's
Duration and Failure metrics can be graphed on a CloudWatch dashboard.

Having a concrete name to the CodeBuild Project allows for the above
mentioned dashboards to reliably graph these across syntheses.