feat(autobump): expose more config options (#337)
feat(autobump): expose more config options

This PR exposes a few additional configuration options in order to support future `merge-back` functionality:

- `branchPrefix` - to use `merge-back` instead of `bump`.
- `branchSha` - to create the branch from `release` instead of `master`
- `labels` - to apply the `pr/no-squash` label on creation.

Note that i've currently done the minimum amount of changes so that we can move forward with the merge back PR automation. 

Its pretty apparent in my opinion that the code has become rather messy, since we are trying to support multiple scenarios using the same construct:

- Bump without a PR (for delivlib release)
- Bump with PR (for jsii and cdk)
- Bump without bump with PR (for merge back).

To that end, I have already started working on refactoring the common functionality out and making it a bit more maintainable and clear. (You can have a look here: https://github.com/awslabs/aws-delivlib/tree/epolon/mergeback)

The merge back PR automation is rather important since we are currently creating the merge-back in such a way that doesn't allow conflict resolution. I feel ok with merging this messy code and changing it when I finish the refactoring (which i am actively working on). 

Just for more context, this is how it will be used in our pipeline eventually:

```typescript
        this.pipeline.autoBump({
          bumpCommand: 'disable',
          versionCommand: `node -p "require('./scripts/get-version')"`,
          pullRequestOptions: {
            labels: ['pr/no-squash'],
            title: 'chore(release): merge release/$VERSION to master',
          },
          branchSha: 'release',
          branchPrefix: 'merge-back'
        });
```

I've also elected to currently not have the merge-back be part of the pipeline itself, but rather simply a project that will be triggered at the end of the MCM (similar to the initial bump).

I will probably integrate it as part of the pipeline along with the refactoring. In the meantime we will be _One Click Away_.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.