fix: change RSA key name algorithm
The RSA key name used to be calculated off of the signing cert's
complete construct path.

This requires it to be replaced if the construct tree is refactored
(and specifically, if the location of the containing stack is moved,
such as it being moved into a `Stage`).

A name that is based off of the stack's deployment name and the
construct path *inside* the stack:

* is indifferent to the stack's location in the construct tree.
* is backwards compatible with simple setups.

For non-backwards compatible cases, the name is now also configurable.