feat: build timeout
Add a build timeout option and default it to the maximum duration
allowed by CodeBuild. Consumers can narrow the timeout window if they
are concerned that builds might get stuck in an infinite loop.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.