feat(bump): skip push & PR if head is behind base
Use `git merge-base --is-ancestor` to determine whether the candidate
`head` is behind the merge `base`, and skip the `git push` step as well
as creating the pull request in that case (GitHub would fail the attempt
to create the PR with HTTP 422 Unprocessable Entity).

This is more reliable then using the `git diff` test, since this would
return differences even if `base` is strictly ahead of `head` (as it
would have changes on top of `base`).

This would make the "auto-merge-back" script idempotent (re-running it
on an already-merged-back release would transparently succeed).