chore(pipeline): refactor failure notification to use bind pattern
The pipeline notifications is now switched to use a bind pattern
with an enum-like class for better APIs.

This makes the API extensible with the ability for users to write
their own if they wish. It also makes the interface cleaner and
reduces property fragmentation on `PipelineProps` for the same
feature, i.e., failure notifications.

BREAKING CHANGES: The property `failureNotifySlack` has been
removed and replaced by `notifyOnFailure()` API.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.