chore: make Nuget and PyPI publishing respect the FOR_REAL variable
As part of investigating how to test publishing of V2, I discovered that neither
the Nuget or PyPI publishing scripts appear to handle the FOR_REAL env variable
that is respected by the other publishing steps. Copying the patterns from the
other publishing steps so these steps can be tested as well.