feat: expose underlying codebuild projects
This PR exposes the underlying `CodeBuild` project created by:

- The `AutoBuild` construct
- The `Shellable` construct

We need this in order to grant various permissions to the project roles.

BREAKING CHANGE: `pipeline.addShellable` now returns the shellable in addition to the action. Use `.action` to retrieve the action 

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.