chore: make Nuget and PyPI publishing respect the FOR_REAL variable
As part of investigating how to test publishing of V2, I discovered that neither
the Nuget or PyPI publishing scripts appear to handle the FOR_REAL env variable
that is respected by the other publishing steps. Copying the patterns from the
other publishing steps so these steps can be tested/mocked as well.

Related changes:
* I had to update the CodeBuild images, as the existing ones were deprecated and couldn't be used to launch new projects.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.