chore: fail NuGet publish step when there are errors
the dotnet nuget push command gets the symbol packages, creates a PUT
request and then a response.

The response was not being parsed and includes a message indicating
"Your package was pushed." when successful. When there are errors,
they are prefixed with `error:`

i.e. error: Response status code does not indicate success: 403 ...

This change checks the log file for errors and fails the publish step.