chore: fail NuGet publish step when there are errors
the dotnet nuget push command gets the symbol packages, creates a PUT
request and then a response. Our pipeline does not currently fail on this command.

added `-o` flag to the `pipefail`. it was previously excluded for idempotency, which
is now being handled by the `--skip-duplicate` flag on the push.

Also added the `--force-english-output` flag which forces the application to run
using an invariant, English-based culture.

docs: https://docs.microsoft.com/en-us/dotnet/core/tools/dotnet-nuget-push

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.