fix: remove repo as required property of AutoMergeBackOptions
The refactoring in #671 unintentionally changed the signature of
AutoMergeBackOptions to make `repo` a required property, and also
added the `stage` property to AutoMergeBackProperties. This change
reverts both back to the pre-refactor state.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.