chore(bump): refactor AutoMergeBackOptions
As discussed in #684, refactoring the AutoMergeBack interfaces to have a more
standard hierarchy. This change removes 'stage' from AutoMergeBackOptions,
and creates a stage-specific AutoMergeBackPipelineOptions.

BREAKING CHANGE: the 'stage' prop of AutoMergeBackOptions has been moved to a
new AutoMergeBackPipelineOptions interface.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.