feat(pipeline): improved pipeline failure metrics
Currently, metrics are emitted for each pipeline failure as
`<pipelineName>_FailedStages`. This metric is emitted per pipeline once
per minute depending on the state of the pipeline at the time.

This means that if a pipeline is in failed state for 20 minutes, there
will be 20 non-zero datapoints for this metric.
This is not very useful in aggregating failure over time and summarizing
them over hours, days or weeks.

This commit changes the approach entirely. Instead of running the
metrics on a fixed schedule, it reacts to CodePipeline events and emits
the appropriate metric. This should result in one non-zero data point
per failure.

This commit also changes the metric name and includes dimensions for
better querying. It now emits a single metric `PipelineActionFailures`
under the namespace `CDK/Delivlib`. This metric has two dimensions -
`Pipeline` with the name of the pipeline and `Action` with the name of
the Action.
This lets higher resolution on the metrics to identify specific Actions
that are failing more often, while letting aggregated metrics.