fix(pypi): cryptography >= 3.4 is not supported by older pip
`twince`, which is used by the python publisher, has a dependency on the `cryptography` module.
As described in their [changelog], starting version 3.4, they require the latest version of `pip`. Otherwise,
the installer will attempt to compile the module, and the Rust compiler will be required.

This fails in the amazon/aws-sam-cli-build-image-python3.6 image since it has an older version of pip installed.

To fix, simply add an instruction to the dockerfile to upgrade to the latest pip version before installing.

[changelog]: https://cryptography.io/en/3.4/changelog.html#v3-4


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.