feat: golang publishing
Add support for publishing golang code from the build artifact (under the `go/` directory) to a GitHub repository.

To add a go publishing step to your pipeline, use `pipeline.publishToGolang()` or create an instance of `PublishToGolang` construct.

See the README of [jsii-release] for more details.

MISC: allow Shellable environment values to be `undefined`, and then they will not be set.

TESTING: updated awslabs/aws-delivlib-sample to include a `go` section in the jsii config and tested that the delivlib test pipeline publishes go bindings to the same repo (under the "golang" branch).

Resolves aws/jsii#2562

[jsii-release]: https://github.com/aws/jsii-release