fix: reuse same repository for multiple tags
The previous implementation for "mirror multiple tags from the same
repository" was broken: it tried to create a repository per repo×tag,
instead of simply creating a repository per repository and mirroring
all the desired tags inside.

A subsequent problem would be that the construct identifiers of the
two sister repos would conflict, and cause problems in our ops codebase
where the repos are created in a separate stack (and hence lead to
Export rename trouble).

Simplify by doing a repo-per-repo and mirroring multiple tags inside
that repo.