feat: make NuGet publisher support yum AND apt-based distributions (#1041)
The new superchain images are based on debian, so we cannot use yum to
install openssl and jq. Detect whether yum or apt-get is installed, and
issue the correct commands. Display a warning if neither of the tools is
found.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.