feat: support release notes for GitHub publishing
Currently, the GitHub publishing job looks for a CHANGELOG file (either in a
default location or set by environment variable), parses the changelog for the
current version, and uses that result as the release notes. This change enables
an alternative approach; creating a RELEASE_NOTES file which contains the
complete release notes, ready for publishing.

The advantage of this new approach might be if you want to do some
post-processing formatting on the changelog (or even include extra notes not
normally present in the default changelog). For CDKv2, this will be used to
combine the changelog entries for the stable release (aws-cdk-lib) and for the
alpha modules.