chore: validate that we are running on a UTC system (#1046)
If not, the output of `node-ical` cannot be trusted, and it's
very confusing when that happens. Until the bug gets fixed upstream,
or we switch libraries, let's just throw a loud and accurate error.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.