fix(s3): S3 publisher incorrectly skips prefix releases
The S3 publisher uses a signaling file in the target bucket to show
that a particular release has already been uploaded, and it should
skip uploading any of the files in the current distribution.

However, it used `aws s3 ls` to do that, which matches prefixes.

So if we previously uploaded `xyz-1.2.3-pre`, and are now trying to
upload `xyz-1.2.3`, it will incorrectly skip the new upload.

Switch to using `aws s3 cp` to fix this.