fix(csr): impossible to produce CSR with keysize > 2KB (#1098)
The Custom Resource used to return the CSR and self-signed certificates
as part of the output payload, but this easily exceeds the 4KB limit
imposed by CloudFormation.

In order to address that, we are now outputting to a specific S3 bucket
instead, and returning pointers to that.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.