fix: ECR mirrorring uses incorrect jsii/superchain tag (#1137)
The default (`latest`) tag was being used, but it no longer exists.
Replaces it with the `1-buster-slim` tag which is the recommended
substitute.

Also, upgrades to Node 14.x where Node 12.x was still being used.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.