fix: missing name in github release (#1141)
Octokit does no default the `name` field to the tag when publishing a
release, which our previous implementation did. Or actually, it
defaulted the tag_name to the name. Now we explicitly pass both `name`
and `tag_name`.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.