chore(deps): upgrade projen (#1142)
Manually bump projen because the automatic dependency upgrades [PR](https://github.com/cdklabs/aws-delivlib/pull/1134) has been failing for a while (we are on it).

This is needed because the new version of projen contains a [fix](https://github.com/projen/projen/pull/1947) to jest issue currently failing our unit tests.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.