fix: Access githubtoken w/ unsafeUnwrap in AutoBuild (#1145)
Directly accessing the secret githubtoken in AutoBuild is no longer allowed. Code must explicitly call unsafeUnwrap() to confirm the developer understands the data value is a secret and the risks associated with accessing it.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.