fix: upgrade nodejs-lts in Windows Shellables
The nodejs version that ships in the standard Windows CodeBuild image
is relatively old (October 2020), so we upgrade the `nodejs-lts` package
in the `install` phase to get something more recent.

This is necessary in order for `jsii@5.0.x` to be install-able & usable,
due to the reliance on the `node:` pseudo-protocol.