chore: use the jsii/superchain image from the ECR Public Gallery (#1258)
Docker Hub throttles requests to pull `jsii/superchain` images, causing the pipeline generated by the integration test to fail. 

Now that we have them also in the ECR Public Gallery, use it in the pipeline.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.