fix: NuGetPublish reads the certificate incorrectly (#1261)
The value of the SSM parameter is the URL of the S3 bucket that contains the certificate (`s3://delivlib-test-x509codesigningkeyrsaprivatekeycert-11xb1jecarfb/self-signed-certificate.pem`), rather than the certificate itself.

Add a step to download the certificate from the bucket, and set up the corresponding policies.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.