feat: support for node18
Change default CodeBuild image to `aws/codebuild/standard:7.0` which has node18.
This version does not ship with yarn, so we
add a check to conditionally install yarn if not present.

Upgrades version of jsii/superchain to a supported one.
Also switch over to use the public ECR for superchain images.
Gives all projects the required permissions to pull from the public ECR.