fix: lock dependency versions of github publish script (#1348)
Adds a package-lock.json file and changes the install command to `npm ci`. Also prints the installed packages for easier debugging.

The existing `yarn.lock` was removed to be consistent with the used package manager.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.