feat: add props on `SignNuGetWithSignerProps` for signer profile and signer owner (#1477)
This PR adds properties on the `SignNuGetWithSignerProps` interface that allow a user to specify the profile and owner that will be used for signing. These properties are added as environment variables that are passed in the payload to the signing lambda as part of the`sign.sh` script. The signing lambda can utilize these values to start the signing job with the specified signer profile and owner. In addition to the above changes, `sign.sh` has been updated to provided the entire path of the dll file to upload to the signing bucket. This fixes an error where the --body value must be a full path.

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.