fix(gpg resource): add passphrase to command line (#1712)
Since the update of the gpg binary, the passphrase now needs to be passed into it (the agent no longer remembers the passphrase from creating the key in the first place).

Needs an additional `--pinentry-mode=loopback` flag to allow the gpg command to honor the `--passsphrase` flag, otherwise it must always pop up a user dialog (and that won't be possible).


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.