fix(gpg resource): survive errors during cleanup (#1713)
When the GPG agent is shutdown, it removes some UNIX sockets from the temporary directory. Their removal may race with the `rmrf` that we run in the custom resource, leading to an `ENOENT` error.

Catch it and continue. Ultimately, this code runs in a Lambda and all files are going to disappear anyway.


-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.