feat: export `PipelineWatcher` (#1742)
`PipelineWatcher` is the construct used in delivlib to monitor CodePipeline pipelines. It is very useful and we can leverage it to monitor pipelines that aren't managed by delivlib (e.g OpsPipeline). 

-----

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.