fix: bump default superchain images (#1799)
* fix: bump default superchain images

delivlib has some places where it defaults to a particular superchain
image, which may grow outdated over time. This came up specifically
in the context of ECR mirroring recently.

Bump the Superchain image we're using to a recent one everywhere
(`bookworm` instead of `bullseye`, a more recent Node because Node 18
is about to be deprecated).

Also make sure that all CodeBuild projects have reasonable descriptions,
because we have hundreds in our account now and we need some way
to tell them apart.

* chore: self mutation

Signed-off-by: github-actions <github-actions@github.com>

* Duplicate superchain information into tests

* Unused import

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>