fix: NuGet job can't write to SSM when using secret role
When the NuGet publishing script assumes a role to retrieve the API key
secret, it pollutes the global environment variables with the role
credentials.

When we then later try to write to SSM, it uses the wrong role (the
secret role instead of the CodeBuild role).

Retrieve the secret credentials in a subshell so they don't interfere
with the CodeBuild credentials.