fix: MirrorSource.fromDir() leads to a looping self-mutation (#1907)
In the CDK Pipeline, we call `fromDir()` with an absolute path and that
path makes its way into the construct ID.

Because in a pipeline the current working directory contains an
arbitrary string, the construt ID will be different every run.

This construct ID makes its way into the asset display name, and the
asset display name makes its way into the Pipeline, which will be
updated on every run and restart on every run.

Give a fixed display name so this doesn't happen anymore. It may
not be unique, but that is not necessary for display names.